fixed texcoord

+++ Shaders/unlit_shader.vert
@@ -16,7 +16,7 @@ layout(location = 3) in vec3 inNormal;
 
 layout(location = 0) out vec4 fragColor;
 layout(location = 1) out vec2 fragTexCoord;
-layout(location = 2) out int fragTextured;
+layout(location = 2) out float fragTextured;
 
 void main() {
     gl_Position = ubo.transform * vec4(inPosition, 1.0);
